[NV INTERNAL]hte: modify gpio controller match function
The linux jammy kernel does not have fwnode member in the
struct gpio_chip. This gives an error when using the upstream
patch for the tegra234 HTE provider. The CL addresses that
concern and implements match function to match the of_node
instead of the fwnnode

Bug 3961133

Change-Id: I542b03e98c263e9cfca7be071cd2a048cd28d948
Signed-off-by: Dipen Patel <dipenp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/canonical/linux-jammy/+/3088882
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>
Tested-by: Gautham Srinivasan <gauthams@nvidia.com>
Reviewed-by: Gautham Srinivasan <gauthams@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Jon Hunter <jonathanh@nvidia.com>