ANDROID: KVM: arm64: Restrict host-to-hyp MMIO donations
Nothing currently prevents the donation of an MMIO region to the
hypervisor for backing e.g. guest stage-2 page-tables, tracing buffers,
hyp vm and vcpu metadata, or any other donation to EL2. However, the
only confirmed use-case for MMIO donations are for protecting the IOMMU
registers as well as for vendor module usage.

Restrict the donation of MMIO regions to these two paths only by
introducing a new helper function.

Bug: 278749606
Bug: 264070847
Change-Id: I914508fb3e3547fcfabca8557bdf7948cb796099
[ Android 15 update: pull in __pkvm_host_donate_hyp_locked() dependency
  from ("ANDROID: KVM: arm64: Donate IOMMU regions to pKVM") ]
Signed-off-by: Quentin Perret <qperret@google.com>