compiler_types.h: Add dummy __nocfi macro
We will add some patches with code that uses this macro to identify
functions to that are not allowed to be subjected to CFI (Control Flow
Integrity), a Clang feature that our kernel does not support.

We do not want to go through the trouble of adding the series of CFI
patches and also we do not want to modify every patch that adds the
__cfi attribute to it's functions. Hence we just add this as an empty
dummy macro.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>