ANDROID: KVM: arm64: per-CPU shared page for pKVM hyp requests
Currently, when the hypervisor needs something from the host, the only
choice is to return an error such as -ENOMEM (toping-up the
vcpu memcache) or to fake a mem_abort for a fault in the case of the
guest memsharing HVC.

This doesn't really scale and prevent from using more detailed
arguments. So let's create a share page which holds a list of
"requests". When the hypervisor returns to the host, the latter just
need to check for the exception ARM_EXCEPTION_HYP_REQ and read the
shared page.

The shared page being still writable by the host, we can't trust it. The
hypervisor is then solely writing it.

Bug: 278749606
Bug: 243642516
Change-Id: I4a7799982e4f1a75d81b26eb3b0eec4c7feb41b0
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>