ANDROID: KVM: arm64: Add unified interface to topup/reclaim hyp allocators
Currently, there is one interface (3 HVCs) to deal with the hyp(heap)
allocator. In preparation to rework the IOMMU memory donation,
we can reuse the same infrastructure as the hyp allocator.
As the IOMMU page table pages are allocated from different allocator,
we need a new abstraction for allocators in the hypervisor that the
HVCs can use.

Bug: 277989609
Bug: 278749606
Change-Id: I7a96e637f535832b3580bfc96aeb670b6a9d1a25
Signed-off-by: Mostafa Saleh <smostafa@google.com>