UPSTREAM: firmware: arm_ffa: Implement the NOTIFICATION_INFO_GET interface
The receiver’s scheduler uses the FFA_NOTIFICATION_INFO_GET interface
to retrieve the list of endpoints that have pending notifications and
must be run. A notification could be signaled by a sender in the secure
world to a VM. The Hypervisor needs to determine which VM and vCPU
(in case a per-vCPU notification is signaled) has a pending notification
in this scenario. It must obtain this information through an invocation
of the FFA_NOTIFICATION_INFO_GET.

Add the implementation of the NOTIFICATION_INFO_GET interface
and prepare to use this to handle the schedule receiver interrupt.
Implementation of handling notifications will be added later.

(cherry picked from commit 3522be48d82b58821442f666eca8b4e547255810)

Bug: 335194469
Bug: 325404073
Bug: 278749606
Link: https://lore.kernel.org/r/20231005-ffa_v1-1_notif-v4-7-cddd3237809c@arm.com
Change-Id: Iff5d6e2de0aa16e4f736083a92940be52954bfe3
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>
Signed-off-by: Sebastian Ene <sebastianene@google.com>