ANDROID: KVM: arm64: Handle block mappings in the pKVM memsharing walker
When a guest share a page with the host, the hypervisor will walk the
guest stage-2 to validate the ownership and read the PA. This walker
will soon encounter block mappings on its way, it'd better be prepared.

Sharing a range within a block mapping will lead to the installation of
a page table. This might require donations. Report -ENOMEM if blocks
will be broken and if not enough memory is found in the memcache.

Bug: 278749606
Bug: 278011447
Change-Id: If8450038ea5b342a300a8902e4ae0dc163df3f12
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>