FROMLIST: arm64: ptdump: Use the mask from the state structure
Printing the descriptor attributes requires accessing a mask which has a
different set of attributes for stage-2. In preparation for adding support
for the stage-2 pagetables dumping, use the mask from the local context
and not from the globally defined pg_level array. Store a pointer to
the pg_level array in the ptdump state structure. This will allow us to
extract the mask which is wrapped in the pg_level array and use it for
descriptor parsing in the note_page.

Bug: 260238318
Bug: 278749606
Link: https://lore.kernel.org/all/20240220151035.327199-4-sebastianene@google.com/
Change-Id: I92d947e54d1583541b959104df1014a922ca4eb4
Signed-off-by: Sebastian Ene <sebastianene@google.com>