ANDROID: drivers/io-pgtable-arm: Don't panic on existing mappings for KVM
Currently, there is a WARN when a map is existing, as this code is
shared with the hypervisor, WARN will cause a panic, previously this
would happen only incase of kernel invalid requests.
However, after supporting idmapped domains, the hypervisor can map
memory multiple times (on teardown we force pages to host even if
they were shared).
Although this can be avoided by checking first if mapping exist, but
this would waste cylces and returning -EEXIST should enough.

Also remove selftest_running from nvhe code as it is not used in the
common code any more.

Bug: 277989609
Bug: 278749606
Change-Id: If4bfdb64951c68e290bd40e6cc6c23de8da377c5
Signed-off-by: Mostafa Saleh <smostafa@google.com>