ANDROID: KVM: arm64: Change the layout of enum pkvm_page_state
The 'concrete' (a.k.a non-meta) page states are currently encoded using
software bits in PTEs. For performance reasons, the abstract
pkvm_page_state enum uses the same bits to encode these states as that
makes conversions from and to PTEs easy.

In order to prepare the ground for moving the 'concrete' state storage
to the hyp vmemmap, re-arrange the enum to use bits 0 and 1 for this
purpose.

No functional changes intended.

Bug: 322736772
Bug: 278749606
Bug: 326193089
Change-Id: Ia6a5a2bd377de0f1336476c7b6539c252861c582
Signed-off-by: Quentin Perret <qperret@google.com>