ANDROID: KVM: arm64: Add completer prot to memory transition
Soon, we need to donate pages from host to hyp that are NC, instead of
having a custom implementation, we can rely on the memory transition
machinery and get all the proper checks.

For that we add the completer prot in pkvm_mem_transition which makes it
identical to pkvm_mem_{share/donation} so the just replace those.

This patch only replace everything with pkvm_mem_transition with no
extra functionality, so functional change intended.

Bug: 277989609
Bug: 278749606

Change-Id: I823b0c897f39856a7d3e50daf2bc1d51b8acce9d
Signed-off-by: Mostafa Saleh <smostafa@google.com>