HACK: cpu_pm: work around non-working cpudile for EL1
This is a test for debugging purposes, not meant for temporary use.

This is a super ugly hack that works around the fact that cpudile doesn't
seem to work when pkvm is enabled. This is probably due to the fact that
cpudile doesn't for some reason work with nVHE mode. So the workaround is
to register a notifier that returns -EBUSY, preventing the
system from entering idle.

This hack is NOT for production use.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>