ANDROID: drivers/arm-smmu-v3-kvm: Fix domain aperture
Use the IAS from the cfg as it can be restricted to be less than the
SMMU.
This also fixes a bug in stage-1 where it was configured to be limited
to VA_BITS while actual IAS where used in aperture.

Instead of keeping copies of variables from the io_pgtable_cfg in the
host SMMU, just copy the whole thing.

Bug: 277989609
Bug: 278749606

Change-Id: I931d445f5bb4108525b6d9673cf4d6abd4946d12
Signed-off-by: Mostafa Saleh <smostafa@google.com>