ANDROID: drivers/arm-smmu-v3-kvm: Fix detach of stage-1
For SMMUs that support S1 and S2, they can transition between S1 and
S2 domains, for that the driver used to reclaim the CD on stage-1
detach. However, that doesn’t work with PASIDs, instead of adding
complex tracking of devices inside the hypervisor, we use the same
approach as STEs, where CDs are never freed, we retain the value of
the CD ptr on detach of S2, and on next attach, that shouldn't cause
issues with S2 domains as this field is ignored with S2 translation.

Bug: 277989609
Bug: 278749606

Change-Id: I8b4027ba26ae3001c251c26b9e3d9191f88f72b4
Signed-off-by: Mostafa Saleh <smostafa@google.com>