ANDROID: KVM: arm64: Huge page support for pKVM guest relax perm
This is the symmetrical operation of wrprotect. Here as well, add
support for huge pages by extending this interface to take order > 0
pages.

We although now split this operation according to its usage. While, the
__pkvm_relax_perm HVC applies on a single page, huge or small and must
be aligned with the actual mapping (similar to wrprotect,
host_unshare_guest and reclaim_page), __pkvm_dirty_logging, applies on a
PAGE_SIZE granule and may break bloc mappings if necessary.

Bug: 278749606
Bug: 278011447
Change-Id: I1739034443979913f04e5de78d0c7690a4725b63
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>