ANDROID: KVM: arm64: Fix printk() overflow for hyp_trace_printk
The cmdline option hyp_trace_printk allows to redirect all tracing
events from the hypervisor directly into dmesg. Sadly, only a check for
the overflow of the seq_buf was in place and the printk limit is way
lower (1024). As we don't have any solution to get this value, which is
meant to be internal only, let's print one event per printk instead.

Bug: 278749606
Change-Id: I68d46b3925fc89dce7f96b3292ece87034a8677c
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>