ANDROID: KVM: arm64: Restore memcache representation
Recently, order was added to memcache, and the idea was to encode the
order in the PAGE_SIZE bits as memcache is always page aligned.
However, due to wrong usage of FIELD_*, now the address is shifted
by PAGE_SIZE, which can cause address corruption under some
configurations.

Also, the naming of the macros make things confusing, just use
PAGE_MASK as it make things more clear.

Bug: 333874255
Bug: 277989609
Bug: 278749606

Change-Id: I2782f8995b7230300b524e2197b607426b2ed250
Signed-off-by: Mostafa Saleh <smostafa@google.com>