ANDROID: KVM: arm64: iommu: Donate domains root
Previously,kvm_iommu_init used to run inside pkvm_init, which means
that creating mapping for hypervisor will be donated from
fix_host_ownership. However, this changed and all the calls were
replaced with donation inside the SMMUv3 driver, but this one was missed.
Convert it as the rest to a donation.

Bug: 277989609
Bug: 278749606

Change-Id: Ib1c621210b2d34cee25eb77975a6288232c63fb0
Signed-off-by: Mostafa Saleh <smostafa@google.com>