ANDROID: drivers/arm-smmu-v3-kvm: Align driver private data
Private data was set to host_arm_smmu_device however, mostly
it was assumed to be arm_smmu_device, this only works because,
arm_smmu_device is the first member of the struct.

Align it to be arm_smmu_device, and use the right accessor to
to make the code more robust.

Bug: 277989609
Bug: 278749606

Change-Id: I07c9282b8c94bc08702829623fb7518404f778d7
Signed-off-by: Mostafa Saleh <smostafa@google.com>