ANDROID: KVM: arm64: iommu: Increase max atomic pool order
atomic pool size was chosen to be of order 6, which was arbitray
choice based on the IOMMU pool.
However, IOMMU pool only holds page table pages, while atomic pool
can be used to implement SMMU emulation, this means it can allocate
much larger size, for example the cmdq on the host is 1MB and STEs
are 4MB.
So, increase the max order of the pool to be of order 10 (4MB).

Bug: 277989609
Bug: 278749606

Change-Id: I68e2812da3ee0b46984d6c67e6565d1f662c2955
Signed-off-by: Mostafa Saleh <smostafa@google.com>