ANDROID: KVM: arm64: iommu: Fix map_pages() error path
commit "a737b7d0e721 (ANDROID: KVM: arm64: iommu: Reduce the logic in
generic code)" moved map operation including io-pgtable loop to the
driver to remove dependcy on io-pgtable from iommu code.
However, the error path was relying on this loop incrementing paddr
so it can properly undo the rest of the range, luckly we know how
much was mapped so we just need to increment it.

Bug: 277989609
Bug: 278749606

Change-Id: I16183590de51a3989c9f8c699ee6513a5effc90d
Signed-off-by: Mostafa Saleh <smostafa@google.com>