ANDROID: KVM: arm64: iommu: Allow driver to flush cached refcount
commit "a737b7d0e721 (ANDROID: KVM: arm64: iommu: Reduce the logic in
generic code)" move unmap_walk logic to the SMMUv3 driver to abstract
the unmap operation.

When that was moved, to avoid exposing the kvm_iommu_flush_unmap_cache()
a WARN_ON was added on the assumption that it is enough to limit the
pgcount called into the driver to KVM_IOMMU_PADDR_CACHE_MAX.
As it turns out, that IOMMU layer in the kernel can map pages with
at page level and unmap them at the block level, so 1 page in a block
expands to 512 enteries in the cache.
So, we need to expose this function to the driver.

Bug: 277989609
Bug: 278749606

Change-Id: Ic3815bdf2f96fd086a7b7c5c5a2ff32478fbf2d3
Signed-off-by: Mostafa Saleh <smostafa@google.com>