ANDROID: KVM: arm64: Fix stage-2 ptdump XN annotation
KVM_PTE_LEAF_ATTR_HI_S2_XN is actually a mask for one of the XN, PXN or
UXN. We do only use XN, so fix that value.

While at it, move that field so it appears after "R W" and invert the
annotation to get the more usual "RWX" output.

Bug: 260238318
Bug: 278749606
Change-Id: I5b69496f22e91d45b5993d97524f298bc97be847
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>