qemu_defconfig: Disable kernel modules that are not needed
The current qemu_defconfig contains a lot of kernel moduels that we do not
need when running in Qemu. There are a lot of modules that support HW that
is not present in Qemu, or features that we do not need.

The reason for doing this is to cut down on build time when building the
host kernel for Qemu.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>