ANDROID: KVM: arm64: Report size correctly from pkvm_mem_abort()
pkvm_mem_abort()'s caller can optionally pass in ipa and size pointers
which are expected to be updated to the address and size of the page
that was faulted in (usually PMD_SIZE or PAGE_SIZE). However, when
pkvm_mem_abort() finds an already faulted in page-size mapping in the
middle of a HP, we fallback to faulting in at page granularity, update
the ipa pointer, but fail to update the size pointer. Fix that.

Bug: 278749606
Bug: 278011447
Change-Id: I672d31e55d5fee6988a8f57a20f76c6fe983653e
Signed-off-by: Quentin Perret <qperret@google.com>