ANDROID: drivers/arm-smmu-v3-kvm: Detach devices on release
When a device is removed from a bus, it calls "release_device"
then frees the device domain if possible.
However, as "kvm_arm_smmu_release_device" doesn't detach the
device from the domain, it's still refcounted and wouldn't be freed.

On release_device, call detach device to decrement the domain refcount
and also puts the device in blocking state.

Bug: 369495618
Change-Id: I91dfdce7ab2ba4b0abc0569b6da081a3456ba2a8
Signed-off-by: Mostafa Saleh <smostafa@google.com>