ANDROID: drivers/arm-smmu-v3-kvm: Setup event queue
The host can use the event queue for debugging, and unlike the command
queue, it would be managed by the kernel.
However, it must set in a shared state so it can't be donated to the
hypervisor later.

This relies on the ARM_SMMU_EVTQ_BASE can't be changed after
de-privilege.

Bug: 277989609
Change-Id: I17cd7ce1a7ce6c9089044f1e95ddcf9b9267ab88
Signed-off-by: Mostafa Saleh <smostafa@google.com>