qemu_defconfig: Add normal tracing options
It may sometimes be useful to trace this or that on the host. For this
reason, we add most the normal tracing options that are enabed by a
Linux distribution like Ubuntu. However, we do not add
CONFIG_IRQSOFF_TRACER, nor CONFIG_PREEMPT_TRACER. These are not normally
enabled because of the large overhead they have, even when not used.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>